LayoutTests/fast/dom/HTMLHtmlElement/html-element-manifest.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>
<script>
description('This tests that manifest IDL attribute does not exist on html element');
shouldBeFalse('"manifest" in document.documentElement');
</script>
</body>
</html>
